check if application version 3
